Start implementing Sqlite3 data provider.
This is quite a big commit because some issues got conflated together:

- some IdentifierQueryExpression types were introduced to aid testing, they identify named components in an SQL query.
- The data provider for Sqlite3 started to get it's rewrite/implementation.
- The design of extension APIs started to take shape.

Extension QueryExpression types don't have any support in the DataKit.SQL directly.
Instead, it's up to the data provider library to implement extension support how the developer sees fit.
The Sqlite3 provider uses a pattern I'd advocate adopting, give the responsibility of writing raw sql
to your extension QueryExpression types and provider a call context that
specifies which dialect of sql you are expressing, in this case by implementing ISqlite3Extension which
takes a specific implementation, Sqlite3QueryWriter, not an abstract QueryWriter.

The Append method that writes raw sql is left to the provider to implement because
adding it to the abstract QueryWriter doesn't provide enough context, the developer
won't know from receiving a QueryWriter object which dialect of sql is being expressed.