Start implementing SQL operators.
The IN operator is implemented, by default, to not wrap SELECT statements (or other executable queries) in parentheses.
In testing with Sqlite3 wrapping an IN-SELECT in parentheses makes the IN operator function against the first
result record from the IN-SELECT only. The default Silk.Data.SQL.Base writer does the same thing and seems to be
passing muster so far but a test of the behavior on each targetted platform should be added to the DataKit project.