Add Parameter support to SQL.
Parameters and values are added to the SQL query expressions.

Values are implemented as parameters and added to the ParameterBag on the QueryWriter. Values without using parameters aren't part of the specification for this library, if bare values are required later it'll be something to carefully consider how to implement and make developers aware of the security implications.