Properly format executable statements with ; and parentheses.
Top-level statements now terminate with a ";" to enable execution of multiple statements in a single query.

Sub-queries are now properly wrapped in parentheses which is in slight contradiction to the note on commit 5e152694c7ae10f957cbe77c84700411e8de02a1. The IN operator still needs to make an exception for sub-queries to avoid a double parentheses formatting in an IN-SELECT situation to avoid the mentioned bug,